requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
shap>=0.43.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
plotly>=5.17.0
scipy>=1.11.0
imbalanced-learn>=0.11.0
folium>=0.15.0
streamlit-folium>=0.15.0
twilio>=8.0.0
python-dotenv>=1.0.0